+++ src/SluggableObserver.php
@@ -46,9 +46,9 @@ class SluggableObserver
     /**
      * @param \Illuminate\Database\Eloquent\Model $model
      * @param string $event
-     * @return boolean|null
+     * @return boolean|void
      */
-    protected function generateSlug(Model $model, $event)
+    protected function generateSlug(Model $model, string $event)
     {
         // If the "slugging" event returns a value, abort
         if ($this->fireSluggingEvent($model, $event) !== null) {
@@ -66,7 +66,7 @@ class SluggableObserver
      * @param  string $event
      * @return mixed
      */
-    protected function fireSluggingEvent(Model $model, $event)
+    protected function fireSluggingEvent(Model $model, string $event)
     {
         return $this->events->until('eloquent.slugging: ' . get_class($model), [$model, $event]);
     }
@@ -76,17 +76,9 @@ class SluggableObserver
      *
      * @param  \Illuminate\Database\Eloquent\Model $model
      * @param  string $status
-     * @return void
      */
-    protected function fireSluggedEvent(Model $model, $status)
+    protected function fireSluggedEvent(Model $model, string $status)
     {
-        if (method_exists($this->events, 'fire')) {
-            // Up to Laravel 5.3, use fire()
-            $this->events->fire('eloquent.slugged: ' . get_class($model), [$model, $status]);
-        } else {
-            // Laravel 5.4+, use dispatch()
-            $this->events->dispatch('eloquent.slugged: ' . get_class($model), [$model, $status]);
-        }
-        
+        $this->events->dispatch('eloquent.slugged: ' . get_class($model), [$model, $status]);
     }
 }